change event counter system to being more safe

+++ src/state/event.go
@@ -26,6 +26,6 @@ type EventInterface interface {
 type Event struct {
 	Type         string
 	Expiration   uint32
-	Counter      uint64
+	Counter      uint32
 	EncodedEvent []byte
 }
